+++ frontend/src/routes/cfp/constants.tsx
@@ -1,0 +1,47 @@
+export const ROW_PADDING = {
+  mobile: 0.5,
+  tabletPortrait: 0.5,
+  tabletLandscape: 0.5,
+  desktop: 1,
+};
+
+export const BUTTON_PADDING = {
+  mobile: 1,
+  tabletPortrait: 1,
+  tabletLandscape: 1,
+  desktop: 2,
+};
+
+export const TOPIC_OPTIONS = [
+  { label: "Select", value: "" },
+  { label: "PyWeb", value: "pyweb" },
+  { label: "Python & Friends", value: "python_e_friends" },
+  { label: "PyData", value: "pydata" },
+];
+
+export const LANGUAGE_OPTIONS = [
+  { label: "English", value: "eng" },
+  { label: "Italian", value: "ita" },
+];
+
+export const TYPE_OPTIONS = [
+  { label: "Select", value: "" },
+  { label: "Talk", value: "talk" },
+  { label: "Training", value: "training" },
+  { label: "Sprint", value: "sprint" },
+];
+
+export const DURATION_OPTIONS = [
+  { label: "Select", value: "" },
+  { label: "2 hours", value: "2h" },
+  { label: "45 minutes", value: "45m" },
+  { label: "30 minutes", value: "30m" },
+  { label: "5 minutes", value: "5m" },
+];
+
+export const AUDIENCE_LEVEL_OPTIONS = [
+  { label: "Select", value: "" },
+  { label: "Beginner", value: "beginner" },
+  { label: "Intermidiate", value: "intermidiate" },
+  { label: "Advanced", value: "advance" },
+];
